+++ packages/veui/src/components/Table/_TableFoot.js
@@ -14,11 +14,14 @@ export default {
     return (
       <tfoot>
         <tr>
-          {this.table.selectable ? <th></th> : null}
           {
-            this.columns.map(col => (
-              <th class={col.align ? `veui-table-column-${col.align}` : null}>{col.renderFoot()}</th>
-            ))
+            this.$slots.default
+              ? <th colspan={this.columns.length + (this.table.selectable ? 1 : 0)}>{this.$slots.default}</th>
+              : (this.table.selectable ? [<th></th>] : []).concat(
+                this.columns.map(col => (
+                  <th class={col.align ? `veui-table-column-${col.align}` : null}>{col.renderFoot()}</th>
+                ))
+              )
           }
         </tr>
       </tfoot>